requirements.txt
# Requires Python >=3.8
requests
python-dotenv
